Add kmp-common-firebase module
