--- appJs/appApiJs/src/commonMain/kotlin/com/gchristov/thecodinglove/Firebase.kt
@@ -1,5 +1,0 @@
-package com.gchristov.thecodinglove
-
-import dev.gitlive.firebase.FirebaseApp
-
-expect fun provideFirebaseApp(): FirebaseApp
